refactor(controllers): extract demo hosts to separate module
- move 7 existing demo host classes from the stories file to
  stories/demo-hosts.ts so they can be imported without polluting CSF
  exports
- add configurable DemoFocusgroupPlayground host that mirrors
  FocusgroupNavigationOptions as reactive properties for Storybook
  controls
- remove excludeStories from meta since demo hosts are no longer
  co-located
